+++ mayan/apps/lock_manager/tests/test_backends.py
@@ -9,9 +9,7 @@ from ..exceptions import LockError
 from .literals import TEST_LOCK_1, TEST_LOCK_LONG_NAME
 
 
-class FileLockTestCase(BaseTestCase):
-    backend_string = 'mayan.apps.lock_manager.backends.file_lock.FileLock'
-
+class BaseLockBackendTestMixin:
     def setUp(self):
         super().setUp()
         self.locking_backend = import_string(dotted_path=self.backend_string)
@@ -80,8 +78,49 @@ class FileLockTestCase(BaseTestCase):
         # The `name` field will be created as a text field with no size limit.
         self.locking_backend.acquire_lock(name=TEST_LOCK_LONG_NAME)
 
+    def test_purge(self):
+        self.locking_backend.acquire_lock(name=TEST_LOCK_1, timeout=30)
 
-class ModelLockTestCase(FileLockTestCase):
+        # lock_1 not release and not expired, should raise LockError
+        with self.assertRaises(expected_exception=LockError):
+            self.locking_backend.acquire_lock(name=TEST_LOCK_1)
+
+        self.locking_backend.purge_locks()
+
+        # lock_1 not release but has expired, should not raise LockError
+        lock_2 = self.locking_backend.acquire_lock(name=TEST_LOCK_1)
+
+        # Cleanup
+        lock_2.release()
+
+
+class DefaultTimeoutTestMixin:
+    @override_settings(LOCK_MANAGER_DEFAULT_LOCK_TIMEOUT=1)
+    def setUp(self):
+        super().setUp()
+        self.locking_backend = import_string(dotted_path=self.backend_string)
+
+    def test_default_timeout_expired(self):
+        self.locking_backend.acquire_lock(name=TEST_LOCK_1, timeout=None)
+
+        # lock_1 not release and not expired, should raise LockError
+        with self.assertRaises(expected_exception=LockError):
+            self.locking_backend.acquire_lock(name=TEST_LOCK_1)
+
+        self._test_delay(seconds=1.01)
+
+        # lock_1 not release but has expired, should not raise LockError
+        lock_2 = self.locking_backend.acquire_lock(name=TEST_LOCK_1)
+
+        # Cleanup
+        lock_2.release()
+
+
+class FileLockBackendTestCase(BaseLockBackendTestMixin, DefaultTimeoutTestMixin, BaseTestCase):
+    backend_string = 'mayan.apps.lock_manager.backends.file_lock.FileLock'
+
+
+class ModelLockBackendTestCase(BaseLockBackendTestMixin, DefaultTimeoutTestMixin, BaseTestCase):
     backend_string = 'mayan.apps.lock_manager.backends.model_lock.ModelLock'
 
 
@@ -89,5 +128,5 @@ class ModelLockTestCase(FileLockTestCase):
 @override_settings(
     LOCK_MANAGER_BACKEND_ARGUMENTS={'redis_url': 'redis://127.0.0.1:6379/0'}
 )
-class RedisLockTestCase(FileLockTestCase):
+class RedisLockBackendTestCase(BaseLockBackendTestMixin, DefaultTimeoutTestMixin, BaseTestCase):
     backend_string = 'mayan.apps.lock_manager.backends.redis_lock.RedisLock'