+++ HISTORY.rst
@@ -256,3 +256,10 @@
 - Added a patch for Django's ``Migration`` class to display time delta for
   each migration during development.
 - The Whoosh search backend is now the default search backend.
+- Lock manager updates:
+
+    - Test the locking backend when the app loads.
+    - Ensure the default timeout setting is used by the ``RedisLock`` backend.
+    - Add support for purging ``RedisLock`` backend locks.
+    - Prefix all locks in the ``RedisLock`` backend to avoid name clashing
+      when using the same database.